ZZALOG: Bug: progress status reports for searching WSJT-X appeared random. Integer overflow was happening calculating percentages. Replaced int by uint64_t in progress data structure. ALL.TXT is on the brink of breaking 32-bit integer in its size.

+++ zzalog/include/status.h
@@ -9,6 +9,7 @@
 #include <fstream>
 #include <list>
 #include <vector>
+#include <cstdint>
 
 using namespace std;
 
@@ -67,12 +68,12 @@ using namespace std;
 
 	// A progress stack entry
 	struct progress_item {
-		int max_value;
-		int value;
+		uint64_t max_value;
+		uint64_t value;
 		char* description;
 		char* suffix;
 		bool countdown;
-		int prev_value;
+		uint64_t prev_value;
 	};
 	
 	// This class provides the status bar and manages access to the various status information
